+++ airflow/kubernetes/worker_configuration.py
@@ -21,7 +21,7 @@ from typing import Dict, List
 import kubernetes.client.models as k8s
 
 from airflow.configuration import conf
-from airflow.kubernetes import pod_refiner
+from airflow.kubernetes import pod_enricher
 from airflow.kubernetes.k8s_model import append_to_pod
 from airflow.kubernetes.pod_generator import PodGenerator
 from airflow.kubernetes.secret import Secret
@@ -396,6 +396,6 @@ class WorkerConfiguration(LoggingMixin):
 
         pod = append_to_pod(pod, self._get_secrets())
 
-        pod = pod_refiner.refine_pod(pod)
+        pod = pod_enricher.refine_pod(pod)
 
         return pod
